An EGL/GLES over Wayland Video Output Sink for Gstreamer ver 0.10
EglGlesWaylandSink renders video frames on a EGL surface it sets up
from a window it creates on Wayland. All the display/surface logic
in this sink uses EGL to interact with the native window system.
The rendering logic, in turn, uses OpenGL ES v2.

This sink has been tested to work on Wayland Desktop Shell
and while it's currently using an slow copy-over rendering
path it has proven to be fast enough on the devices we have tried it on.